Remplacé les types 'long' par des 'int' pour les egids et les ewids.
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@55724 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/business-it/src/test/java/ch/vd/uniregctb/interfaces/ServiceCivilTest.java
@@ -105,8 +105,8 @@ public class ServiceCivilTest extends BusinessItTest {
 
 		for (Adresse adresse : adresses) {
 			if (adresse.getTypeAdresse() == TypeAdresseCivil.PRINCIPALE) {
-				assertEquals(Long.valueOf(3037134), adresse.getEgid());
-				assertEquals(Long.valueOf(3), adresse.getEwid());
+				assertEquals(Integer.valueOf(3037134), adresse.getEgid());
+				assertEquals(Integer.valueOf(3), adresse.getEwid());
 			}
 		}
 	}
